feature/IT-30 Refactor for repository location folders
