+++ src/main/java/com/tourism/assetmanagement/repository/asset/AssetSportRepository.java
@@ -1,4 +1,4 @@
-package com.tourism.assetmanagement.repository;
+package com.tourism.assetmanagement.repository.asset;
 
 import com.tourism.assetmanagement.domain.AssetSport;
 import org.springframework.data.jpa.repository.Query;
