+++ src/main/java/com/tourism/assetmanagement/repository/asset/BaseAssetObjectRepository.java
@@ -1,4 +1,4 @@
-package com.tourism.assetmanagement.repository;
+package com.tourism.assetmanagement.repository.asset;
 
 import com.tourism.repository.BaseRepository;
 import org.springframework.data.repository.NoRepositoryBean;
