+++ src/main/java/com/tourism/assetmanagement/utils/AssetClassificationUtil.java
@@ -6,8 +6,8 @@ import com.tourism.assetmanagement.domain.classification.*;
 import com.tourism.assetmanagement.errors.NotFoundException;
 import com.tourism.assetmanagement.errors.ServiceException;
 import com.tourism.assetmanagement.mapper.AssetClassificationMapper;
-import com.tourism.assetmanagement.repository.AssetClassificationRepository;
-import com.tourism.assetmanagement.repository.AssetManifestationRepository;
+import com.tourism.assetmanagement.repository.asset.AssetClassificationRepository;
+import com.tourism.assetmanagement.repository.asset.AssetManifestationRepository;
 import com.tourism.assetmanagement.repository.ManifestationRepository;
 import com.tourism.assetmanagement.repository.classification.*;
 import com.tourism.domain.BaseEntity;
