+++ src/main/java/com/tourism/assetmanagement/utils/GenericUtil.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.tourism.assetmanagement.errors.NotFoundException;
 import com.tourism.assetmanagement.errors.ServiceException;
-import com.tourism.assetmanagement.repository.BaseAssetObjectRepository;
+import com.tourism.assetmanagement.repository.asset.BaseAssetObjectRepository;
 import com.tourism.domain.BaseEntity;
 import com.tourism.repository.BaseRepository;
 import org.springframework.util.CollectionUtils;
