+++ src/main/java/com/tourism/assetmanagement/utils/RouteUtil.java
@@ -3,9 +3,9 @@ package com.tourism.assetmanagement.utils;
 import com.tourism.assetmanagement.domain.AssetRoute;
 import com.tourism.assetmanagement.domain.Route;
 import com.tourism.assetmanagement.domain.RouteType;
-import com.tourism.assetmanagement.repository.AssetRouteRepository;
+import com.tourism.assetmanagement.repository.asset.AssetRouteRepository;
 import com.tourism.assetmanagement.repository.RouteRepository;
-import com.tourism.assetmanagement.repository.RouteTypeRepository;
+import com.tourism.assetmanagement.repository.type.RouteTypeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
